Moved get_owner_url
